requirements.txt
asyncio
PySocks
telethon
opentele
2captcha-python